chore: rename navigation panel
