+++ frontend/src/components/NavigationPanel.vue
